Re-structured ('smoothing' to handle zero counts)
Re-structuring is a work in progress. Implementing Laplace smoothing. To do: (1) When calculating probabilities, iterate through the map with all bigram combinations entered (currently the code iterates over the training text bigrams again to see if the map contains them, which shouldn't be a concern anymore; all bigrams entered should be considered). (2) Work with that map instead of creating a new probability map (obvious memory implications). (3) Do not add 1 again to those that were the 'zero-count' bigrams (i.e. those that already had been 'smoothed' / given a weight of 1 when entered into the map).